+++ TerminalFormatter/Compatibility/StoreRotationConfig.cs
@@ -1,0 +1,41 @@
+using GameNetcodeStuff;
+using HarmonyLib;
+using MrovLib.ContentType;
+using StoreRotationConfig.Api;
+
+namespace TerminalFormatter
+{
+  internal class StoreRotationConfigCompatibility : MrovLib.Compatibility.CompatibilityBase
+  {
+    public StoreRotationConfigCompatibility(string guid, string version = null)
+      : base(guid, version)
+    {
+      if (IsModPresent)
+      {
+        Init();
+      }
+    }
+
+    public static void Init()
+    {
+      if (!Plugin.isSRCPresent) // In case anything calls `Init()` again for some reason...
+      {
+        UnpatchTerminalScroll();
+      }
+
+      Plugin.isSRCPresent = true;
+    }
+
+    public static int GetDiscountedPrice(BuyableThing buyable, out int discount)
+    {
+      return RotationSalesAPI.GetDiscountedPrice(buyable.Nodes.Node, out discount);
+    }
+
+    public static void UnpatchTerminalScroll()
+    {
+      // Prioritize TerminalFormatter's scrolling settings.
+      Plugin.harmony.Unpatch(AccessTools.Method(typeof(PlayerControllerB), "ScrollMouse_performed"),
+        HarmonyPatchType.Transpiler, "pacoito.StoreRotationConfig");
+    }
+  }
+}