+++ TerminalFormatter/Nodes/Store.cs
@@ -289,7 +289,108 @@ namespace TerminalFormatter.Nodes
           continue;
         }
 
-        table.AddRow(decorationName, $"${buyable.Price}", "");
+        // StoreRotationConfig compatibility.
+        if (Plugin.isSRCPresent)
+        {
+          int price = StoreRotationConfigCompatibility.GetDiscountedPrice(buyable, out int discount);
+
+          if (discount > 0)
+          {
+            string discountPercent = $" {(decor ? '(' : "")}-{discount}%{(decor ? ')' : "")}";
+
+            if (decorationName.Length + discountPercent.Length > Settings.itemNameWidth)
+            {
+              decorationName = decorationName[..(Settings.itemNameWidth - 4 - discountPercent.Length)] + "..." + discountPercent;
+            }
+            else
+            {
+              decorationName = $"{decorationName.PadRight(Settings.itemNameWidth - discountPercent.Length)}{discountPercent}"
+                .PadRight(Settings.itemNameWidth);
+            }
+          }
+
+          table.AddRow(decorationName, $"${price}", "");
+        }
+        else
+        {
+          table.AddRow(decorationName, $"${buyable.Price}", "");
+        }
+        // ...
+
+        if (ConfigManager.DivideShopPage.Value != 0)
+        {
+          if (itemCount % ConfigManager.DivideShopPage.Value == 0)
+          {
+            itemCount = 1;
+            if (ConfigManager.ShowGroupDividerLines.Value)
+            {
+              table.AddRow("".PadRight(Settings.itemNameWidth, '-'), "".PadRight(5, '-'), "".PadRight(5, '-'));
+            }
+            else
+            {
+              table.AddRow("", "", "");
+            }
+          }
+          else
+          {
+            itemCount++;
+          }
+        }
+      }
+
+      table.AddRow("", "", "");
+      table.AddRow("[SUITS]", "", "");
+      itemCount = 1;
+      if (decor)
+      {
+        table.AddRow($"{new string('-', Settings.dividerLength)}", "", "");
+      }
+
+      List<BuyableSuit> SuitSelection = ContentManager.Suits.OrderBy(x => x.Name)
+        .Where(x => terminal.ShipDecorSelection.Contains(x.Nodes.Node)).ToList(); // TODO: 'InRotation' property for BuyableSuit, or a parent class.
+
+      foreach (var buyable in SuitSelection)
+      {
+        UnlockableItem unlockable = buyable.Suit;
+        string suitName = buyable.Name;
+
+        if (decor)
+        {
+          suitName = $"* {suitName}";
+        }
+
+        if (unlockable.hasBeenUnlockedByPlayer || unlockable.alreadyUnlocked)
+        {
+          continue;
+        }
+
+        // StoreRotationConfig compatibility.
+        if (Plugin.isSRCPresent)
+        {
+          int price = StoreRotationConfigCompatibility.GetDiscountedPrice(buyable, out int discount);
+
+          if (discount > 0)
+          {
+            string discountPercent = $" {(decor ? '(' : "")}-{discount}%{(decor ? ')' : "")}";
+
+            if (suitName.Length + discountPercent.Length > Settings.itemNameWidth)
+            {
+              suitName = suitName[..(Settings.itemNameWidth - 4 - discountPercent.Length)] + "..." + discountPercent;
+            }
+            else
+            {
+              suitName = $"{suitName.PadRight(Settings.itemNameWidth - discountPercent.Length)}{discountPercent}"
+                .PadRight(Settings.itemNameWidth);
+            }
+          }
+
+          table.AddRow(suitName, $"${price}", "");
+        }
+        else
+        {
+          table.AddRow(suitName, $"${buyable.Price}", "");
+        }
+        // ...
 
         if (ConfigManager.DivideShopPage.Value != 0)
         {